refactor(phase2): manual_generation と manual_editing を manual/ に統合
