+++ lib/features/manual/data/services/gemini_image_service.dart
