+++ lib/features/manual/data/services/image_annotation_service.dart
