+++ lib/features/manual/data/services/image_extraction_service.dart
