+++ lib/features/manual/data/services/video_analysis_service.dart
@@ -1,11 +1,11 @@
 import '../../../../core/utils/result.dart';
-import '../../../video_upload/domain/entities/video_file.dart';
-import '../entities/manual_step.dart';
+import '../../../video/domain/entities/video_file.dart';
+import '../../domain/entities/manual_step.dart';
 
-/// Interface for Gemini API service
-abstract class GeminiService {
+/// Interface for video analysis service
+abstract class VideoAnalysisService {
   /// Analyzes a video file and extracts manual steps
-  /// 
+  ///
   /// Requirements: 2.1, 2.2, 2.3, 2.4
   /// - 2.1: Video is sent to Gemini API when upload completes
   /// - 2.2: Maximum 20 steps are extracted
