+++ lib/features/manual/domain/entities/manual_step.dart
