+++ lib/features/manual/domain/value_objects/manual_generation_progress_stage.dart
