--- lib/features/manual_editing/data/data.dart
@@ -1,2 +1,0 @@
-// Data layer exports for manual editing feature
-export 'services/manual_edit_service_impl.dart';
