--- lib/features/manual_editing/domain/domain.dart
@@ -1,2 +1,0 @@
-// Domain layer exports for manual editing feature
-export 'services/manual_edit_service.dart';
