--- lib/features/manual_generation/data/data.dart
@@ -1,7 +1,0 @@
-// Manual Generation Data Layer Exports
-export 'repositories/manual_repository_impl.dart';
-export 'services/gemini_service.dart';
-export 'services/gemini_image_service.dart';
-export 'services/image_extraction_service.dart';
-export 'services/video_analysis_service.dart';
-export 'services/pdf_export_service_impl.dart';
